% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{editable.submissionsync-package}
\alias{editable.submissionsync}
\alias{editable.submissionsync-package}
\title{editable.submissionsync: Interactive Excel-Style Data Editor}
\description{
A lightweight table viewer package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ramdhadage/editable.submissionsync}
  \item Report bugs at \url{https://github.com/Ramdhadage/editable.submissionsync/issues}
}

}
\author{
\strong{Maintainer}: Ramnath Dhadage \email{ram.dhadage123@gmail.com}

}
\keyword{internal}
