% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{get_schema_config}
\alias{get_schema_config}
\title{Get Schema Configuration}
\usage{
get_schema_config(config = Sys.getenv("GOLEM_CONFIG_ACTIVE", "default"))
}
\arguments{
\item{config}{Environment name}
}
\value{
List with schema configuration
}
\description{
Retrieves schema configuration from golem-config.yml.
}
\examples{
\dontrun{
schema_config <- get_schema_config()
}
}
