% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton_loader.R
\name{skeleton_content}
\alias{skeleton_content}
\title{Skeleton Content Placeholder}
\usage{
skeleton_content(id)
}
\value{
HTML showing skeleton placeholder
}
\description{
Shows immediately while Handsontable and actual data initialize
}
\keyword{internal}
